llama-model/requirements.txt
# requirements.txt

# Web framework
fastapi
uvicorn

# Hugging Face API and local models
transformers
torch
sentencepiece
huggingface-hub
requests 

# Configuration
python-dotenv
